household_simulator/requirements.txt
fastapi
uvicorn[standard]
python-socketio[asyncio_client]
tensorflow
numpy